feat(normalization): Optionally mark scrubbed transactions as sanitized (#1917)
If the corresponding feature flag is set, mark transactions that were
scrubbed by regex patterns as `sanitized` instead of `url`.

### Motivation

If we discovered identifiers in a URL transaction by pattern matching,
we previously did not risk marking it as `sanitized`, because it could
theoretically still contain user names or other high-cardinality
identifiers. But in practice this seems to conservative, and we'd rather
assume that a URL scrubbed by pattern matching is low-cardinality
enough.

### Risks

If there is a second identifier in the URL that neither the
pattern-matching nor the clusterer discovered, we mark something
high-cardinality as low-cardinality. This might add noise to the product
and an increase in transaction metrics being rate-limited by the
cardinality limiter.

https://github.com/getsentry/team-ingest/issues/69