feat(mobile): Move device.class from contexts to tags (#1911)
`device.class` is currently being calculated and stored in the `device`
context. This change moves the calculation and storage into tags, under
the `device.class` key. This is because we need `device.class` to be
surfaced by various tag endpoints in the Sentry product, and we don't
need `device.class` in contexts.