ref(actix): Refactor shutdown without actix (#1912)
Removes a dependency on `actix` from Relay's shutdown functionality.
Instead, this uses tokio directly to get notified about process signals.
The `Controller` no longer creates an actix system as a result of this,
which also means that it is now reduced to a service that simply handles
shutdowns.

The Controller's API is purely static. This makes it easier to use the
controller from various places across the code base. This is subject to
be revised in a near-term followup.

Since the web server still requires an actix system, this is now handled
by the `HttpServer` service internals. It creates a new thread and
blocks on the actix system inside of this thread. The shutdown message
is then forwarded into that thread. This required to make startup of the
`HttpServer` asynchronous, as we need to wait for the service to start
in order to obtain a handle to the inner actix server.