ref(profiling): Remove platform validation (#1933)
Let's align with the event protocol and not validate platform for
profiles.

`runtime` is not a required field, the profile is valid even if those
fields are omitted. I'd remove the validation for `cocoa` as well but we
do have some code relying on these fields for Snuba ingestion so we'll
need to deal with that first in order to treat everything as optional
like for the event.