fix(quota): New reason code for fallback rate limit (#1967)
When a non-processing Relay receives a 429 response when sending
envelopes to the upstream, it logs a `RateLimited` outcome with a reason
code that is parsed from the `X-Sentry-Rate-Limits` header. However, if
that header is not available, it logs an outcome with
`RateLimited(None)`. This happens for example when Sentry SaaS rejects
the envelope on the nginx layer.

Because reason code `None` is hard to debug, replace it with a dedicated
reason code that we can search for.

Follow-up work: Modify the nginx anti-abuse layer to send an
`X-Sentry-Rate-Limits` header.

Fixes https://github.com/getsentry/relay/issues/1929.