ref(tx-processor): Apply transaction rules after UUID scrubbing (#1964)
Applying transaction rules before UUID scrubbing makes it impossible to
track what rules were applied. It's interesting to know what rules are
useful to reduce the high cardinality, so that we can use them again
over other rules that are less interesting or provide no value. Applying
these rules after UUID scrubbing leaves no room for further modification
of the transaction name, so the resulting transaction name is guaranteed
to be the direct output of a rule.

This is the first step in the process of discarding the rules that
provide no value and prioritizing those that do.