build(auth): Update cryptography dependencies (#1975)
This updates `ed25519-dalek`, `hmac`, `rand`, and `sha2` to their
respective latest versions. The Ed25519 crate is currently in final
development for version 2. We're using this version as this allows us to
run the latest versions of all other crates.

This allows us to remove the `failure` dependency completely from our
dependency tree.