build: Remove time from chrono dependencices (#1977)
By default, `chrono` pulls an outdated version of `time` that has a
known security vulnerability. Chrono is not affected by this
vulnerability, but the dependency still creates an alert in our
dependency tree.

This PR disables default features in all of our own crates and updates
all dependencies that relied on time v1.