fix(feat-flags): Fix serialization of unknown feature flags (#2040)
Currently, unknown feature flags that end up in the `Unknown` bucket are
serialized as `"Unknown"`, so the actual feature flags don't reach
downstream relays. Since `serde` doesn't yet support types with
`"other"` fields, I've manually implemented the required serialization
and deserialization methods.

As a result of this change, feature flags that have been removed in
relay and are no longer required will still be sent to downstream
relays. This is a trade-off we must accept to address the issue above,
and this new issue should be resolved by not adding the feature flags to
project configs in Sentry in the first place.