ref(replays): increase chunking limit to 15MB (#2032)
We were previously chunking kafka messages on the replay recording topic
to 1 MiB. We are aiming to eliminate this chunking as it requires us to
maintain a redis cluster / constrains us to in-order processing. Other
infra at Sentry has successfully experimented with larger kafka
messages, so we feel this is safe.

With https://github.com/getsentry/relay/pull/2031 this should pretty
much mean we no longer chunk any messages. We can remove the chunking
logic once this is confirmed after deploy.

- [x] Increase chunking limit from 1MiB to 15 MiB. 

We chose 15MiB as the limit, as we're eventually going to combine the
replay event with the replay recording, and will want some extra
headroom for this in addition to our 10 MiB recording limit.

Depends on ops increasing our kafka payload size limit first.