ref(spooler): Add state into the buffer service (#2055)
This is an initial refactor of the `BufferService` with introduced
`BufferState`. If the on-disk spool file configuration option provided
the state is switched between:
* `MemoryFileStandby` - use memory for buffering and when the max limit
reached, switch to `File` and continue spooling to and reading from disk
* `MemoryFileRead` - the file size hit the limit, and we switched to
spooling to in-memory buffer only, but still continue to read from disk,
when disk is empty we switch state to `MemoryFileStandby` or to `File`,
if the there is more space for in-memory buffer to fit onto the disk
* `File` - use disk for spooling to and reading from, when the disk is
full we switch state to `MemoryFileRead`

If there is no on-disk spooling, only 1 state `Memory` will be always
used.

This separation makes possible to better reason about what parts of the
system currently involved, where the spool is located. Avoiding
unnecessary accesses to the system, which is not used at the moment
(e.g. if the in-memory buffer used, in normal circumstances there is no
need to hit the disk all the time).

In terms of functionality almost nothing should change, except as
mentioned above, depending on the state some parts of the buffer /
spooler won't be used till the required state is reached, and I only
changed test to accommodate this (see test `metrics_work `). This is
mostly about the organisation , introducing helper structures and
shuffling code around.


### State changes can be represented as following: 

```mermaid
%%{init: {'theme':'forest'}}%%
stateDiagram-v2
direction LR

MemoryFileStandby --> File : Memory is full 
File --> MemoryFileStandby : Disk is empty
File --> MemoryFileRead : Disk is full
MemoryFileRead --> File : Disk has enough \n space for in-memory buffer
MemoryFileRead --> MemoryFileStandby: Disk is empty

```