fix(quota): Parse large limits (#2079)
Rate limits for attachments are measured in bytes, but Relay cannot
parse rate limits that do not fit into a `u32`. This means that the
highest possible rate limit is 4 gibibytes (GiB), which became a problem
in [INC-362](https://getsentry.atlassian.net/browse/INC-362).

Change the internal datatype to `u64` to support larger numbers for rate
limits.