ref(crons): Remove monitor slugification
For now we would prefer to simply limit the length and let the
validation of monitor slugs happen at the consumer side.

This way we do not need to try and match the slugification logic that
python uses.

With the slugification happening in relay it becomes very opaque why a
check-in my not be getting through.