instr(buffer): Count number of envelopes (#2093)
Bring back envelope counts as statsd metrics for the spooling envelope
buffer:

1. `envelopes_mem_count`: Number of envelopes currently in memory.
2. `envelopes_written`, `envelopes_read`: Number of envelopes flushed to
and from disk in any given time frame.
3. `envelopes_disk_count`: Number of envelopes on disk: Because running
a `SELECT count(*)` query on a large table is expensive, we _only_ track
this number if there is nothing in the database on startup. This should
be the default use case, so in general, this metric will be available.
If there is data in the db on startup, only `envelopes_written` /
`envelopes_read` will be available.

**Discarded options for `envelopes_disk_count`**

* We considered running the `SELECT count(*)` query async on startup,
but that might still lock the database table.
* We considered running the counting query in a `tokio::timeout`, but it
is unclear if this would actually unlock the database, or only cancel
the future while the query keeps running in the background.
* We could limit the max. number of rows in the counting query, such
that the metric still works for small leftovers in the db, but not for
large buffers encountered on start up. This should be easy to implement
in a follow-up if the need for it occurs.
* We could get track this number in the db itself (in a separate table),
and update it every time we spool or unspool, from within a database
transaction. We decided this is too much engineering effort for just a
metric, but might reconsider if it turns out we really need this metric
in the edge case that there's something in the db on startup.

Fixes https://github.com/getsentry/team-ingest/issues/115