fix(buffer): Throttle dequeue (#2096)
There is a maximum limit on in-memory envelopes in Relay
(`envelope_buffer_size`). Whenever the spooler loads envelopes from
disk, we consume some of this budget. If we dequeue too eagerly, there's
a chance that we exhaust the budget, which will lead to Relay dropping
new envelopes coming in via web requests.

* To prevent this, wait until there is a `>= 50%` budget of permits
available before dequeueing a batch from disk.