feat(spans): Use table name as domain in sql-like descriptions (#2142)
The domain should represent the table name in SQL-like span
descriptions. This PR adds this capability by extracting the table name
with a RegEx. Note that this is a best effort and the RegEx may not
cover all queries; this is ok for now, we'll improve in further
iterations as needs evolve.