feat(spans): Apply span description rules (#2161)
Applies span description name rules to span descriptions.

Transaction name rules also apply to span descriptions for now, but this
will soon not be the case. As a result, the avoidable clones in the
transaction processor will also go away (this is a shortcut for now). On
the other hand, we are still applying rules to `op=http*` spans,
ignoring the rule's scope; this is another shortcut that will get
addressed at a later time.