doc: document pii fields (#1972)
A tool which takes the path of a rust crate or workspace and outputs all
the fields of structs or enums with the metaattribute 'pii = "true"' (or
custom values), with the full path. E.g.
`Event.csp.effective_directive`.

The current use case will be to get all the pii=true fields under the
Event struct, in order to display it in sentry-docs.

if needed, in the future it can be trivially generalized to look for any
custom metaattribute key, not just "pii".