feat(txname): Mark all URL transactions as sanitized (#2210)
In https://github.com/getsentry/relay/pull/2139, we transitioned to
marking all URL transactions as "sanitized" once sentry provides the
"ready" flag. However, this results in a bad onboarding experience for
new projects because they still see `<< unparameterized >>` for the
first ~10 hours.

This PR removes the "ready" condition, such that _all_ transactions with
source URL will now get their transaction name added as a metric tag.
Since we were already doing this for all but the newest projects, this
should not add much cardinality after all. But we have to keep an eye on
the cardinality limiter in the first few hours / days after deploying
this.

Fixes https://github.com/getsentry/relay/issues/2186

---------

Co-authored-by: Iker Barriocanal <32816711+iker-barriocanal@users.noreply.github.com>