feat(spans): Forward already scrubbed SQL queries (#2211)
In span metrics, we only add the span description tag if we managed to
scrub the span description. In database spans, it's common for SDKs to
already send scrubbed queries, and usually Relay won't do further
scrubbing, resulting in the span description not being present. With
this PR, Relay will check if the description looks like it has already
been scrubbed, and extract the span description in these cases.

#skip-changelog