fix(docs): Fix filtering of "other"-fields (#2222)
There is a special case with the fields that have the attribute
`additional_properties`, where we want to pass them through only if they
also have the attribute `retain = "true"`. To support this, I refactored
the tool, so that now it collects all of the attributes of all the
matching pii-fields, and then filters it based on this logic.


Also, the output is now reverted to json, and the output schema has a
boolean for whether the field has `additional_properties` or not.

Due to us now collecting all the attributes at the stage before the
output, it makes this tool easier to use if we wanna document some other
attribute in the future.