ref(spans): Introduce types for span metrics and tags (#2224)
Span metrics and types were initially introduced as a hard-coded way of
directly typing strings. To avoid mistyping in the future, this PR
introduces types so that there's no need to write strings. The approach
it's pretty much the same as with transaction metrics, but a bit simpler
due to less strict requirements.