feat(geo): Add GeoIp lookup to light normalization (#2229)
This change introduces the GeoIp lookup to light normalization process,
which allows us to populate the `User` with correct geo information as
soon as we can. Which also allow to set a correct tag in metrics
extraction , if the metrics extraction is happening on our side (in the
PoPs in this case).

I still keep the GeoIp lookup in the store normalization, since it's
also used in few places, more specifically in `relay-cabi`, which
exposed few different functions to the python lib.

Lookup won't be triggered in any case if the Geo information is either
provided by user or set on any of the steps.