feat(txname): Apply transaction rules to legacy SDKs (#2250)
https://github.com/getsentry/sentry/pull/51437 expanded the scope of
transaction name clustering to transactions from old SDKs that do not
send a `source` annotation yet. This PR enables the application of rules
to this type of transactions in Relay.

Because there is a potential of triggering the cardinality limiter, I
gated this new behavior with a feature flag.

The condition for applying clusterer rules has now become more
complicated, so I decided to remove / ignore the `scope` configuration
flag entirely.