feat(metrics): Implement generic metric extraction for transactions (#2257)
Adds an initial version of generic metrics extraction based on the
previously added schema (#2252). It is limited to transaction events and
evaluates conditions with the fields supported by `FieldValueProvider`.

Since generic metric extraction can operate on all item types, the a new
top-level processing function is added that runs before the more
specific transaction metric extraction. In the future, we will convert
conditional tagging logic and even transaction metric extraction
configuration to the new format when the project config is read, so that
the old functions can be removed.

Since the new function supports only transactions, it reuses the
existing `transaction_metrics_extracted` flag from the processing state.
This is subject to change in a follow-up.