fix(spans): Run SQL scrubber even if already parameterized (#2286)
In https://github.com/getsentry/relay/pull/2211, we fixed the problem
that noop-scrubbing led to an empty span description:

> In database spans, it's common for SDKs to already send scrubbed
queries, and usually Relay won't do further scrubbing, resulting in the
span description not being present.

But that change had the side effect that queries partially scrubbed on
the client side were not scrubbed any further in relay. With this PR, we
restore best-effort scrubbing on those, for two reasons:

* Catch additional parameters that the SDK did not mark as such.
* Better aggregation by replacing `IN (%s, %s, %s)` (variable list
length) with `IN (%s)`.