ref(server): Change internal flagging for metric extraction (#2297)
Restructures the signatures of the metric extraction functions so that
there are distinct top-level functions that can be called and tested
independently of each other. These functions are:

- generic metric extraction
- transaction metric extraction
- conditional tagging for transaction metrics
- span metric extraction

All of these functions are called from a single place `extract_metrics`
in the envelope processor. The transaction metrics flag is checked a
single time.