fix(test): Increase timeout of each iteration for `BufferGuard` enter (#2326)
I notice it's usually fails when it tries to enter the `BufferGuard`,
meaning we do not have any permits left at this point.
In this test it can happen only if we are not fast enough to spool the
incoming envelopes out of the memory to the disk.

Increasing the timeout for the iteration in the for loop, should help to
avoid this, when e.g. actions runners get somewhat slow.