ref(server): Internal Arc for EnvelopeProccesorService (#2363)
This refactor is needed for #2320, made it a separate PR to make it
easier to read the diffs on that branch

We want to make envelopeprocessorservice easily cloneable, so we put
most of the data it contains behind an arc. This allows us to remove the
outer arc in the spawn_handler, so that we can mutate the service. Which
will be needed for global config update.

#skip-changelog