feat(spans): Extract spans from transactions (#2350)
Extract spans from transactions and publish them on a new Kafka topic
`ingest-spans`. Protected by a feature flag.

Notes:

* Spans are extracted only in processing relays. If an SDK or downstream
Relay provides `Span` items in an envelope, they are deleted
unconditionally.
* The extraction happens after all regular transaction processing, for
example metrics extraction and dynamic sampling. If we want to split the
sampling logic in the future, we will have to move the extraction to an
earlier stage.