ref(metrics): Move span metric config to project config (#2398)
This PR adds a top-level function to extract metrics from all event
types. This comprises transaction metrics as well as span metrics.
Internally, it delegates most of the heavy-lifting to the generic metric
extraction function. The intent is to create one top-level extraction
function for every envelope item type that is capable of traversing data
and instantiating `FieldValueProvider`s.

All event-specific logic is therefore moved to the events module. Most
importantly, this includes the implementations of `Extractable`, but
also the logic for traversing through spans.

The configuration for span metric extraction is now located in the
project config (`relay-dynamic-config`). It is added automatically if
the project has the span extraction feature enabled.