chore(test): Deflake the actors::spooler::tests::dequeue_waits_for_permits (#2404)
Make sure we are waiting and consuming sent messages to ensure 
that slow Github action agents are also covered. 

This test does not require testing of the timeouts or timings, 
but only the proper flow of the messages in the service which will set the correct internal state 
and will produce correct results.