ref(metrics): Validate transaction timestamps in light normalization (#2406)
As opposed to errors, transactions have a valid timestamp range
defaulting to (-5d..+1h). If transactions with an end timestamp outside
of this range are ingested, they are dropped as invalid transactions.

The reason for this check are performance metrics: Sentry metrics cannot
be backdated more than five days, and normalizing the transaction
timestamp like for errors creates strange artifacts in metrics.

However, this check was originally implemented in metric extraction
code. In this PR, it is moved to light normalization, which already
performs a series of validations of transaction events.