ref(protocol): Move `FieldValueProvider` to a new `Getter` trait (#2437)
The `FieldValueProvider` trait was originally introduced to abstract
data access for dynamic sampling rules. Since then, the same rules are
being used for conditional metric tagging and metric extraction.

Since accessing fields in data instances by path is a more generic
concern, the trait is now moved to `relay_protocol` and renamed to
`Getter`. It retains its single `get_value` method, but will be extended
by more functionality for arrays and objects in the near future.

To simplify the implementation of the trait, the `get_value` method now
returns an option. This option is `None` if either the path does not
match any known field, or if the field is an empty option.

To avoid redundant allocations, `get_value` no longer returns an owned
`serde_json::Value`. Instead, there is a new `Val` type: The borrowed
variant of `Value`. For now, the only data actually borrowed are
strings, however, there are stubs to support arrays and objects.