ref(sampling): Split the sampling crate into submodules (#2438)
Refactor that splits the relay-sampling crate into several submodules.
The entire crate was implemented and tested in a single file, which
exceeded a reasonable size.

As part of this refactor, missing documentation is added to types that
were previously undocumented.

Most of the tests still reside in the top-level library source file,
since they test e2e sampling functionality. If the API surface of the
crate is changed, these tests can also be moved to a better location.