fix(globalconfig): Shutdown the service on Relay shutdown
When Relay gets a SIGTERM and network access is disabled for
production pods, the global config service continues to request configs
from upstream until Relay gets a SIGINT or SIGKILL. As a result, the
service can't make network requests and emits error metrics, which can
be confused from actual errors.

This PR introduces the shutdown control in the service, shutting it down
with Relay and avoiding the unnecessary network requests and error
metrics.