fix(server): Do not call shutdown on runtimes (#2477)
This change should prevent panics on shutdown, that is

> A Tokio 1.x context was found, but it is being shutdown.

From the [tokio
docs](https://docs.rs/tokio/latest/tokio/runtime/struct.Runtime.html#shutdown):

> Tasks spawned through
[Runtime::spawn](https://docs.rs/tokio/latest/tokio/runtime/struct.Runtime.html#method.spawn)
keep running until they yield. Then they are dropped.

> The
[shutdown_background](https://docs.rs/tokio/latest/tokio/runtime/struct.Runtime.html#method.shutdown_background)
and
[shutdown_timeout](https://docs.rs/tokio/latest/tokio/runtime/struct.Runtime.html#method.shutdown_timeout)
methods can be used if waiting forever is undesired. When the timeout is
reached, spawned work that did not stop in time and threads running it
are leaked.

In our case, waiting for _blocking_ tasks is acceptable, as long as they
stop when they yield.