fix(server): Immediatly set global config to watch contents. (#2505)
Envelopeprocessorservice updated the global config only when the watch
changed, however, if we use static config then the watch won't change
after we first set its contents to the static config. If we subscribe
after it was first set, we wouldn't ever get to use the static config in
the envelope processor service.

to fix this, we immediatly update the global config in envelope
processor service to the contents of the watch.