feat(spans): Update db table extraction to add comma separators
While we work on adding support for keys with multiple values in metric
tags, in the short term we want to unblock work in the database module.
To do this, we're going to perform a wildcard search on the metric value
in the backend on a comma seperated list.

The `span.domain` is currently emitted as a comma seperated list when
there are multiple table values, but no commas are added if there single
table values. This patch updates that to always add commas.

For example what was the `users` table would become `,users,`. This is
important because the wildcard search going to be added is going to do
something similar to `*,table_name,*`. We don't want to confuse between
two tables like `org` and `org_users`, which can happen when we don't
have the wildcard.

This is a hack, but given we are changing this anyway when we get multiple
I think it's fine to add as it's needed to unblock the db module GA.