feat(spans): Allow activerecord with db.system (#2545)
We've seen SQL query normalization fail when the queries are prepended
with an active record comment, for example

```SQL
/*some comment `my_function'*/ SELECT 1
```

Normalization works fine when `sqlparser` is able to parse the query
though. Having a `db.system` in `span.data` is a strong indicator that
we will be able to parse the query. So we re-allow activerecord spans if
they have a `db.system` set.

In addition, this PR also improves the fallback regex behavior for this
kind of comments.