ref(sampling): Remove dependency from sampling to event schema (#2548)
The `relay-sampling` crate implements generic functionality to define
and evaluate dynamic sampling rules. Originally, the functions were tied
to the event schema, but now can run independently of that based on the
`Getter` trait.

This PR removes the build dependency on `relay-event-schema`. The only
remaining use was in a constructor of `DynamicSamplingContext` that
creates the DSC from a transaction event. Since this is business logic
specific to the envelope processor, this utility is now moved to
`relay_server::utils::dynamic_sampling`.