ref(sampling): Refactor sampling evaluator (#2557)
A follow-up to #2550. 

The main motivation for this PR is that the `SamplingRule::Evaluate`
method became pretty ugly because it relied on that the two variants of
`SamplingValue` both had a sample rate attached, which it no longer does
with the introduction of the `Reservoir` variant.

The function has mainly been replaced by
`SamplingEvaluator::try_compute_sample_rate`. Whose job it is to
validate the rule, and if it's valid, return a sample rate.

Validating for the `SamplingRate` and `Factor` variant means that it's
not out of bounds with the given time range (depending on the decaying
function), and for the `Reservoir` variant it means that the limit has
not been exceeded.

It will return an optional `ControlFlow`, where `None` means the rule is
invalid and should be skipped, `Break` is analogous to
`SamplingValue::SampleRate` (but includes reservoir), and `Continue` is
analogous to `SamplingValue::Factor`.

Some other adjustments have also been made, this should be all the
changes:

* Created the `try_compute_sample_rate` method as described.
* Moved the decaying function logic to its own method.
* Moved the checking of time range constraints to the beginning of for
loop, so we don't do the more expensive condition-matching first.
* Updated tests



Co-authored-by: Jan Michael Auer <mail@jauer.org>