ref(server): Add a dedicated metric for transaction usage (#2571)
Adds a dedicated and cheap counter metric for usage tracking. It
contains the `has_profile` tag, which will later be removed from the
durations metric to lower its cardinality.

This metric should improve the performance of all usage queries. In the
long term, it allows to decouple product metrics from bookkeeping.