ref(spool): Clearer error message on unsupported paths
Sqlx seems to not support relative paths. In those cases, the spool file
can't be read and the actor fails to set up. This PR attempts to provide
more clarity into the error message.

By default, the spool file is created if it doesn't exist, but it's
assumed the directory to exist, failing if that assumption is not valid.
This PR also attempts to create the parent directories recursively if
they don't exist, so that the spool file can successfully be created.