feat(feedback): add User Feedbacks as Events (#2604)
Relay changes for the revamped User Feedback feature. User Feedbacks
(Known as `UserFeedbackV2`), will be events that flow through our system
instead of attachments. See
[here](https://www.notion.so/sentry/User-Feedback-Beta-11d68db039bc436bae017ad56c0c20f9?pvs=4)
for more information on the tech spec (and the link at the bottom for
what we do once they're ingested).


Note that within this code, we're calling it `UserReportV2`. This is to
avoid confusion with the existing `UserReport`. A PR to shim the
previous `UserReports` as new `UserReportV2`'s will follow, at which
point we'll rename this in the code to be `UserFeedback`. The string
`feedback` is used for the context and ItemType / EventType parsing.

- [x] Adds a feature flag for ingest `UserReportV2Ingest`. This is
backed by options https://github.com/getsentry/getsentry/issues/11573
- [x] Creates a new EventType `UserReportV2`
- [x] Creates a new ItemType `UserReportV2`
- [x] Creates a new DataCategory `UserReportV2`
- [x] Adds a new context `UserReportV2Context` that for now contains two
properties, `contact_email` (no pii stripping), and the `message` which
contains the feedback text.

---------

Co-authored-by: Iker Barriocanal <32816711+iker-barriocanal@users.noreply.github.com>