ref(normalization): Introduce NormalizeProcessor (#2668)
We're working on having a single processor to run all
normalization-related steps in Relay, and this PR introduces it:
`NormalizeProcessor`. This is only the first step towards that goal, and
more PRs will follow up to merge other processors. The scope of this PR
is the following:

- Create a new `NormalizeProcessor`.
- Since there was already a processor with such a name in
`StoreProcessor`, rename this one to `StoreNormalizeProcessor`. Note
that this will eventually be merged into `NormalizeProcessor` too.
- `StoreNormalization` has no further modifications. It'll be addressed
in future PRs.
- Move light normalization steps into `NormalizeProcessor`. These are
moved directly, copy-pasting, without additional refactors.
- Keep the old `light_normalize_event` function. This function is called
from plenty of places in the codebase including tests, so we'll keep it
until `NormalizeProcessor` is completed and only then remove the
function indirection, in order to minimize the risk of bugs.

There are no functional changes in this PR.