Update README.md (#2680)
## Description:

This Pull Request fixes a typographical error in the README.md file.
#skip-changelog


## Changes Made:

 - Corrected "avalibale" to "available" in the README.md file.

## Additional Information:

This fix is not related to any existing issue. It's a minor typo that I
noticed while reviewing the README.

Thank you for considering this contribution.


### Legal Boilerplate

Look, I get it. The entity doing business as "Sentry" was incorporated
in the State of Delaware in 2015 as Functional Software, Inc. and is
gonna need some rights from me in order to utilize my contributions in
this here PR. So here's the deal: I retain all rights, title and
interest in and to my contributions, and by keeping this boilerplate
intact I confirm that Sentry can use, modify, copy, and redistribute my
contributions, under Sentry's choice of terms.