ref(metrics): Flatten AggregatorServiceConfig (#2681)
Prior to splitting the aggregatorservice with the aggregator we had just
one big config. After we split it, we also split the config to reflect
the service and hte aggregator, where the service config was just the
aggregatorconfig flattened together with the values related to the
service.

This nested structure, while it makes more sense in the implementation,
made it harder to configure. This PR therefore 'unflattens' the service
config so it looks as it did in the past. We keep the config for the
inner aggregator and simply implement a `From<&aggregatorserviceconfig>
for aggregatorconfig` in order to keep some of the implementation
benefits of separation.
 
The only downside is that it's less 'dry'. The values for inner
aggregator is now present in both the serviecconfig and the inner
config, so thats something we have to keep in mind in the future if we
add new values.