chore(metric): Add metrics to track the project state type (#2695)
Add the `full` tag to `ProjectStateRequest` metric, to measure what
types of the project states are requested.

Also add the `ProjectStateResponse` metric with the same tag. This is mainly will be used to track the responses by processing relays to PoPs in order to confirm that the PoPs should be getting full project configs.