feat(schema): Add Getter for fields commonly used in tags (#2707)
Sentry extracts the `logger` and `level` fields into tags, and both of
these fields are commonly used in dashboard widgets. For this reason,
on-demand metric extraction needs access to those fields.

Sentry will convert queries for the tags into the corresponding fields,
as the tag does not exist in Relay yet.