ref(normalization): Reduce processor module visibility (#2729)
The are a lot of tests requiring `NormalizeProcessor` in
`transactions/processor.rs`, forcing higher visibility than necessary on
`normalize/processor.rs`. This PR moves required tests from one module
to another, thus reducing the required module visibility. As we continue
to refactor the processors, we'll find a better home for functions and
tests.