feat(server): Delay serving global config in managed mode before upstream response (#2697)
## Problem
currently, before we request global config from upstream, we will simply
serve the default global config. This kind of implicit behaviour is not
ideal.


## Solution
This PR will affect the behaviour in managed mode, where we won't start
processing envelopes until we have received a valid global config. To do
this we hook into the existing buffer logic in projectcache where we
buffer until we receive a projectstate for an envelope, and simply add
global config as a second requirement.

## Implementation

* we move the logic for getting global config from the envelopeprocessor
to projectcache, and we include it in the message we send to
envelopeprocessor for processing, which guarantees there's a valid
global config at time of processing.
* When we receive an envelope, we check if we have a global config
`handle_validate_envelope`, if not we buffer the envelope.
* When we receive projectstates, this used to always trigger a dequeue
from the buffer, but now if we don't have global config we stop dequeing
and push the projectkey that would have been dequeued into a BTreeSet.
* When we receive a global config, we request new projectstates for all
of the projectkeys in the BTreeSet, which will trigger a new dequeue
when their projectstates arrive, this time it should dequeue
successfully as we have a global config.

The reason we don't dequeue directly when we receive the global config
is because the projectstates might have expired by the time the global
config arrives. This might lead to some redundant fetching of
projectstates but i figured it would be rather insignificant and it will
only happen once, the added complexity of checking was not worth it imo.
If you disagree, let me know!